feat: modularize remaining packet builders
